Added option to show custom attributes or not
